+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog for Websauna
 1.0a12 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Replace deprecated logger.warn with logger.warning.
 
 
 1.0a11 (2018-10-10)